Add use of sequence number switch
When dealing with an unknown label, creating a new label or adding a
label to the label file, ask for a sequence number and pass it to
DisPass. Only the `dispass2' algorithm uses this, but the `dispass1'
algorithm ignores it, so it shouldn't be a problem to just always send
it. Eventually the capabilities of each algorithm should be gathered
from DisPass itself.