Add dispass-insert
Does the same thing as the `dispass' command, except it inserts the
results into the current buffer, instead of copying them into the
clipboard. This is useful when Emacs asks for a password and thus skips
the clipboard and `kill-ring'.

Using the universal argument is not possible because the prefix argument
is already used to determine the required length of the generated
passphrase.